CS/requirements.txt
# Core Dependencies
# Note: TensorFlow 2.13.0 requires numpy<=1.24.3
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Deep Learning
tensorflow==2.13.0
keras==2.13.1

# Explainability
shap==0.42.1

# Web Framework & Dashboard
flask==2.3.2
flask-cors==4.0.0
plotly==5.17.0

# Data Processing
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3

# Utilities
tqdm==4.66.1
joblib==1.3.2
requests==2.31.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Documentation
sphinx==7.1.2

# IP Geolocation (for location features)
geoip2==4.7.0

# Type hints (compatible with TensorFlow 2.13.0)
# TensorFlow 2.13.0 requires typing-extensions<4.6.0,>=3.6.6
typing-extensions==4.5.0

